requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
google-cloud-api-keys>=0.8.0,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0,<7.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.27.0,<1.0.0
